pyproject.toml
[tool.poetry]
name = "chatapp"
version = "0.1.0"
description = ""
authors = ["MujyKun <mujy@irenebot.com>"]

[tool.poetry.dependencies]
python = "^3.9"
black = "22.6.0"
quart-openapi = {version = "1.7.2", extras = ["doc"]}
python-dotenv = "0.19.2"
quart = "0.17.0"
Werkzeug = "2.0.2"
hypercorn = "^0.13.2"
aiohttp = "^3.8.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
